Move TableStylized component to src
